Extract lenient platform classes from `NodeState`
This is just a refactoring of the virtual platform metadata code,
so that classes related to virtual edges are now top-level instead
of nested into `NodeState`.
